requirements.txt
asgiref==3.2.7
astroid==2.4.1
boto3==1.14.7
botocore==1.17.7
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.6
django-allauth==0.42.0
django-cleanup==5.0.0
django-formtools==2.2
django-localflavor==3.0.1
django-qrcode==0.3
django-social-share==2.0.0
django-storages==1.9.1
django-widget-tweaks==1.4.8
docutils==0.15.2
easy-thumbnails==2.7
google-api-core==1.21.0
google-api-python-client==1.10.0
google-auth==1.19.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauth2client==4.1.3
oauthlib==3.1.0
opensimplex==0.3
Pillow==7.1.2
protobuf==3.12.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.5.2
python-dateutil==2.8.1
python-stdnum==1.13
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.1.0
wrapt==1.12.1
